Nodes Ramp, Worker

+++ include/types.hpp
@@ -4,7 +4,7 @@
 using ElementID = unsigned int;
 using size_type = unsigned int;
 using ProbabilityGenerator = double;
-using TimeOffset=double;
+using TimeOffset=unsigned int;
 using Time= unsigned int;
 // TODO: add Time, TimeOffset i ProbabilityGenerator
 #endif //NETSIM_TYPES_HPP
